Split rewrite tests into a dedicated @Suite
The 7 new tests in 1d4f43a pushed BuildSystemTests' struct body to
1025 lines, just past SwiftLint's 1000-line type_body_length cap.
Move them verbatim into ResourceBundleRewriteTests to keep both
suites under the cap.
